added template functionality for adding multiple service locations
